✨ #6 履歴を使用するためにモデル初期化を別ファイルでやる
